Add MariaDB Optimization Script
This commit introduces a Bash script to optimize MariaDB installations on Debian-based systems. The script works by:

- Checking if the script is run as root
- Checking if the script has been run before
- Updating package lists and upgrading packages
- Installing MariaDB if it's not already installed
- Prompting the user to select an optimization level (either to use up to 50% of total RAM and CPU, or up to 100% if MariaDB is the sole service on the server)
- Backing up the original MariaDB configuration file
- Calculating hardware-based settings (based on total memory and CPU cores)
- Configuring MariaDB performance settings based on the chosen optimization level and hardware
- Restarting the MariaDB service if the user agrees

Please note that the script requires root permissions to run. It also creates a flag file at `/var/run/script_flag` to indicate that it has been run before. Backups of the MariaDB configuration file are stored with the current timestamp to prevent overwriting of previous backups.